pyproject.toml
[tool.poetry]
name = "flasklogmonitor"
version = "0.1.0"
description = ""
authors = ["cainky <kylecain0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.6"
flask = "^2.3.2"
requests = "^2.31.0"
flask-testing = "^0.8.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
